Memory estimation (#27)
* Return struct instead of interface

* bump ggcr

* uses gguf_parser for memory estimation

* Updates all model overviews

+++ tools/model-cards-cli/internal/domain/model.go
@@ -13,7 +13,7 @@ type ModelVariant struct {
 	Quantization  string
 	Size          uint64
 	ContextLength uint32
-	VRAM          float64
+	VRAM          uint64
 	Descriptor    types.ModelDescriptor
 }
